src/{config.ts,popup.ts} - add in basic fields and experimental Popup API
The API is an attempt to encapsulate all the functions a custom Popup
might need to do its thing, which is currently a cut-and-paste job for
all projects.  And to do them a bit more rigorously - specifically,
HTML/URL escape text which might not be trusted.

The fields in the pop-up are fairly standard ones for typical Mukomap
projects, but not really the complete list required for
workers.coop. Will add those in a next step, now the basic API is
working.